+++ src/enums.ts
@@ -1,0 +1,11 @@
+export enum Align {
+  BottomLeft = 'bottom-left',
+  BottomRight = 'bottom-right',
+  TopLeft = 'top-left',
+  TopRight = 'top-right',
+}
+
+export enum Format {
+  Banner = 'banner',
+  Box = 'box',
+}