+++ src/templates/miniGDPR.ts
@@ -1,20 +1,22 @@
-import type { AMGDPR } from '@/elements/AMGDPR'
+import type AMCookies from '@/elements/AMCookies'
 import icon from '@/templates/icon'
 
 /**
  * Mini GDPR
  */
-export default function miniGDPR(this: AMGDPR) {
+export default function miniGDPR(this: AMCookies) {
   if (!this.gdprContainer) {
     return
   }
   this.gdprContainer.innerHTML = /* HTML */ `<button
     class="mini-gdpr ${this.alignMiniPrompt}"
     data-hide="false"
-    aria-label="${this.text?.miniGDPR || 'Cookie settings'}"
+    aria-label=""
   >
     <figure class="icon-cookies settings">${icon}</figure>
-  </button> `
+  </button>`
+
+  this.populateText()
 
   const button = this.gdprContainer.querySelector('.mini-gdpr')
   if (button instanceof HTMLButtonElement) {