+++ types/index.d.ts
@@ -1,3 +1,3 @@
-import { AMGDPR } from './elements/AMGDPR';
+import AMGDPR from './elements/AMGDPR';
 export default AMGDPR;
 export declare const tagName = "am-gdpr";