+++ types/templates/popUp.d.ts
@@ -1,2 +1,2 @@
-import type { AMGDPR } from '../elements/AMGDPR';
+import type AMGDPR from '../elements/AMGDPR';
 export default function popUp(this: AMGDPR): void;