+++ types/templates/switchButton.d.ts
@@ -1,4 +1,4 @@
-import type { AMGDPR } from '../elements/AMGDPR';
+import type AMGDPR from '../elements/AMGDPR';
 export default function switchButton(this: AMGDPR, { disabled, label, name, value, }: {
     disabled?: boolean;
     name: string;